pairwise site phylogenetic community dissimilarity (PCD) within a community

Calculate pairwise site PCD, users can specify expected values from pcd_pred().

pcd(comm, tree, expectation = NULL, cpp = TRUE, verbose = TRUE, ...)

Arguments

comm
A site by species data frame or matrix, sites as rows.

tree
A phylogeny for species.

expectation
nsp_pool, psv_bar, psv_pool, and nsr calculated from pcd_pred().

cpp
Whether to use loops written with c++, default is TRUE.

verbose
Do you want to see the progress?

...
Other arguments.

Value

A list of a variety of pairwise dissimilarities.

References

Ives, A. R., & Helmus, M. R. 2010. Phylogenetic metrics of community similarity. The American Naturalist, 176(5), E128-E142.

Examples

x1 = pcd_pred(comm_1 = comm_a, comm_2 = comm_b, tree = phylotree, reps = 100)
pcd(comm = comm_a, tree = phylotree, expectation = x1)

#> Dropping tips from the phylogeny that are not in the community data

#> 1 2 3 4 5 6 7 8 9 10 11 12 13 14 

#> $PCD
#>           s3293     s3294     s3295     s3296     s3297     s3299     s3300
#> s3294 0.9386582                                                            
#> s3295 1.1165949 0.7075770                                                  
#> s3296 1.1010318 1.3004318 1.2889086                                        
#> s3297 1.1057317 1.1684367 1.1282893 0.7043345                              
#> s3299 0.9542690 0.3430500 0.7278677 1.4089613 1.3986494                    
#> s3300 1.1215818 0.5785085 0.5247930 1.3588246 0.8334648 0.4959597          
#> s3301 0.7124159 0.9882495 1.2964916 1.2479825 1.2549372 1.0006099 1.3032248
#> s3302 0.8609658 0.3805125 1.0678397 1.1054040 1.0725756 0.5481041 0.8733149
#> s3303 0.6615528 0.7404306 1.0867417 0.8756031 0.8343026 0.9040524 1.1403204
#> s3304 0.8438061 0.8826240 1.1982947 0.6398605 0.5688327 1.0323657 1.2536405
#> s3307 0.9328841 0.3483447 0.6650961 1.3686082 1.3777321 0.3743544 0.8171385
#> s3308 0.9246295 0.6239530 1.2559919 0.8215847 0.6331837 0.9575590 1.1236258
#> s3309 1.1179223 0.7772768 0.7721443 1.2912366 1.2799135 0.3949196 0.5256742
#> s3311 1.1162547 0.6622612 0.6620836 1.1339984 0.6253449 0.9455213 0.4703425
#>           s3301     s3302     s3303     s3304     s3307     s3308     s3309
#> s3294                                                                      
#> s3295                                                                      
#> s3296                                                                      
#> s3297                                                                      
#> s3299                                                                      
#> s3300                                                                      
#> s3301                                                                      
#> s3302 0.9065587                                                            
#> s3303 0.7306250 0.4277444                                                  
#> s3304 0.8588853 0.6249252 0.2804423                                        
#> s3307 0.9907211 0.5494079 0.8908461 1.0149040                              
#> s3308 0.9730303 0.4929828 0.5031090 0.2739384 0.9452508                    
#> s3309 1.2474209 0.6734513 1.1908729 1.2598778 0.7773619 1.0982763          
#> s3311 1.2849301 0.8500859 0.9186025 1.0217344 0.9324877 1.1037188 1.0556730
#> 
#> $PCDc
#>           s3293     s3294     s3295     s3296     s3297     s3299     s3300
#> s3294 0.7346939                                                            
#> s3295 1.1052632 0.6521739                                                  
#> s3296 1.0975610 1.3278689 1.3636364                                        
#> s3297 1.0975610 1.3278689 1.0909091 0.6428571                              
#> s3299 0.7346939 0.2769231 0.6521739 1.3278689 1.3278689                    
#> s3300 1.1052632 0.6521739 0.5000000 1.3636364 0.8181818 0.3913043          
#> s3301 0.3600000 0.8490566 1.2000000 1.1739130 1.1739130 0.8490566 1.2000000
#> s3302 0.6585366 0.4426230 1.0909091 0.9642857 0.9642857 0.4426230 0.8181818
#> s3303 0.3600000 0.8490566 1.2000000 0.7826087 0.7826087 0.8490566 1.2000000
#> s3304 0.5454545 0.9473684 1.2857143 0.5294118 0.5294118 0.9473684 1.2857143
#> s3307 0.7346939 0.2769231 0.6521739 1.3278689 1.3278689 0.2769231 0.6521739
#> s3308 0.7346939 0.5538462 1.1739130 0.7377049 0.7377049 0.8307692 1.1739130
#> s3309 1.0909091 0.6315789 0.7142857 1.2352941 1.2352941 0.3157895 0.4285714
#> s3311 1.1020408 0.8307692 0.6521739 1.0327869 0.7377049 0.8307692 0.6521739
#>           s3301     s3302     s3303     s3304     s3307     s3308     s3309
#> s3294                                                                      
#> s3295                                                                      
#> s3296                                                                      
#> s3297                                                                      
#> s3299                                                                      
#> s3300                                                                      
#> s3301                                                                      
#> s3302 0.7826087                                                            
#> s3303 0.5625000 0.3913043                                                  
#> s3304 0.6923077 0.5294118 0.2307692                                        
#> s3307 0.8490566 0.4426230 0.8490566 0.9473684                              
#> s3308 0.8490566 0.4426230 0.5094340 0.3157895 0.8307692                    
#> s3309 1.1538462 0.5294118 1.1538462 1.2000000 0.6315789 0.9473684          
#> s3311 1.1886792 0.7377049 0.8490566 0.9473684 0.8307692 1.1076923 0.9473684
#> 
#> $PCDp
#>           s3293     s3294     s3295     s3296     s3297     s3299     s3300
#> s3294 1.2776181                                                            
#> s3295 1.0102525 1.0849515                                                  
#> s3296 1.0031623 0.9793375 0.9451996                                        
#> s3297 1.0074444 0.8799338 1.0342652 1.0956314                              
#> s3299 1.2988662 1.2387918 1.1160639 1.0610696 1.0533039                    
#> s3300 1.0147645 0.8870463 1.0495859 0.9964714 1.0186791 1.2674525          
#> s3301 1.9789331 1.1639383 1.0804097 1.0630962 1.0690206 1.1784961 1.0860207
#> s3302 1.3073925 0.8596763 0.9788530 1.1463449 1.1123007 1.2383093 1.0673848
#> s3303 1.8376466 0.8720627 0.9056181 1.1188262 1.0660534 1.0647729 0.9502670
#> s3304 1.5469779 0.9316586 0.9320070 1.2086253 1.0744617 1.0897194 0.9750537
#> s3307 1.2697589 1.2579114 1.0198140 1.0306802 1.0375513 1.3518354 1.2529457
#> s3308 1.2585235 1.1265817 1.0699190 1.1137037 0.8583157 1.1526173 0.9571627
#> s3309 1.0247621 1.2306883 1.0810021 1.0452867 1.0361204 1.2505788 1.2265731
#> s3311 1.0128978 0.7971663 1.0151949 1.0979984 0.8476897 1.1381274 0.7211919
#>           s3301     s3302     s3303     s3304     s3307     s3308     s3309
#> s3294                                                                      
#> s3295                                                                      
#> s3296                                                                      
#> s3297                                                                      
#> s3299                                                                      
#> s3300                                                                      
#> s3301                                                                      
#> s3302 1.1583806                                                            
#> s3303 1.2988889 1.0931246                                                  
#> s3304 1.2406121 1.1804142 1.2152501                                        
#> s3307 1.1668493 1.2412548 1.0492187 1.0712876                              
#> s3308 1.1460134 1.1137759 0.9875842 0.8674714 1.1378019                    
#> s3309 1.0810981 1.2720747 1.0320898 1.0498981 1.2308231 1.1592917          
#> s3311 1.0809730 1.1523387 1.0819097 1.0784975 1.1224388 0.9964128 1.1143215
#> 
#> $D_pairwise
#>           s3293     s3294     s3295     s3296     s3297     s3299     s3300
#> s3294 0.8306670                                                            
#> s3295 0.9857201 0.4186715                                                  
#> s3296 0.9798550 0.8653963 0.8273000                                        
#> s3297 0.9840376 0.7775577 0.7242047 0.4929596                              
#> s3299 0.8444818 0.2135267 0.4306774 0.9376193 0.9307570                    
#> s3300 0.9901225 0.3423020 0.2911608 0.8721764 0.5349684 0.2934580          
#> s3301 0.6416168 0.7594899 0.9839312 0.9791681 0.9846247 0.7689891 0.9890411
#> s3302 0.7662101 0.2532190 0.6854045 0.7736659 0.7506895 0.3647460 0.5605466
#> s3303 0.5958084 0.5690360 0.8247482 0.6869989 0.6545945 0.6947827 0.8654100
#> s3304 0.7538969 0.6260509 0.8298862 0.4703337 0.4181243 0.7322636 0.8682163
#> s3307 0.8255572 0.2168223 0.3935356 0.9107655 0.9168372 0.2330117 0.4834987
#> s3308 0.8182523 0.3883709 0.7431671 0.5467387 0.4213638 0.5960193 0.6648464
#> s3309 0.9988055 0.5513274 0.5347532 0.9491321 0.9408090 0.2801190 0.3640588
#> s3311 0.9878312 0.4122153 0.3917532 0.7546401 0.4161473 0.5885266 0.2783005
#>           s3301     s3302     s3303     s3304     s3307     s3308     s3309
#> s3294                                                                      
#> s3295                                                                      
#> s3296                                                                      
#> s3297                                                                      
#> s3299                                                                      
#> s3300                                                                      
#> s3301                                                                      
#> s3302 0.7112867                                                            
#> s3303 0.6041933 0.3356086                                                  
#> s3304 0.6877471 0.4593554 0.2245625                                        
#> s3307 0.7613893 0.3656136 0.6846334 0.7198779                              
#> s3308 0.7477936 0.3280645 0.3866495 0.1943062 0.5883583                    
#> s3309 0.9988646 0.4950249 0.9535842 0.9593070 0.5513878 0.7790144          
#> s3311 0.9874950 0.5657053 0.7059648 0.7247228 0.5804140 0.6869945 0.7487957
#> 
#> $dsor_pairwise
#>           s3293     s3294     s3295     s3296     s3297     s3299     s3300
#> s3294 0.6666667                                                            
#> s3295 1.0000000 0.4545455                                                  
#> s3296 1.0000000 1.0000000 1.0000000                                        
#> s3297 1.0000000 1.0000000 0.8000000 0.5000000                              
#> s3299 0.6666667 0.2000000 0.4545455 1.0000000 1.0000000                    
#> s3300 1.0000000 0.4545455 0.3333333 1.0000000 0.6000000 0.2727273          
#> s3301 0.3333333 0.7142857 1.0000000 1.0000000 1.0000000 0.7142857 1.0000000
#> s3302 0.6000000 0.3333333 0.8000000 0.7500000 0.7500000 0.3333333 0.6000000
#> s3303 0.3333333 0.7142857 1.0000000 0.6666667 0.6666667 0.7142857 1.0000000
#> s3304 0.5000000 0.7500000 1.0000000 0.4285714 0.4285714 0.7500000 1.0000000
#> s3307 0.6666667 0.2000000 0.4545455 1.0000000 1.0000000 0.2000000 0.4545455
#> s3308 0.6666667 0.4000000 0.8181818 0.5555556 0.5555556 0.6000000 0.8181818
#> s3309 1.0000000 0.5000000 0.5555556 1.0000000 1.0000000 0.2500000 0.3333333
#> s3311 1.0000000 0.6000000 0.4545455 0.7777778 0.5555556 0.6000000 0.4545455
#>           s3301     s3302     s3303     s3304     s3307     s3308     s3309
#> s3294                                                                      
#> s3295                                                                      
#> s3296                                                                      
#> s3297                                                                      
#> s3299                                                                      
#> s3300                                                                      
#> s3301                                                                      
#> s3302 0.6666667                                                            
#> s3303 0.5000000 0.3333333                                                  
#> s3304 0.6000000 0.4285714 0.2000000                                        
#> s3307 0.7142857 0.3333333 0.7142857 0.7500000                              
#> s3308 0.7142857 0.3333333 0.4285714 0.2500000 0.6000000                    
#> s3309 1.0000000 0.4285714 1.0000000 1.0000000 0.5000000 0.7500000          
#> s3311 1.0000000 0.5555556 0.7142857 0.7500000 0.6000000 0.8000000 0.7500000
#>